refactor: move scanning functionality to scan.rs
